changed manage category API from service to category

+++ spheroo-admin/src/main/java/com/qburst/spherooadmin/category/CategoryServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -38,13 +39,21 @@ public class CategoryServiceImpl implements CategoryService{
 
     @Override
     @Transactional
-    public void updateCategoryById(Long categoryId, Category category) {
+    public boolean updateCategoryById(Long categoryId, Category category) {
 //        categoryRepository.updateCategoryById(category.getCategoryName(), category.getCategoryDescription(), category.getCategoryIcon(), categoryId);
-        categoryRepository.save(category);
-        List<Long> noReferenceChargeIds = serviceChargeRepository.findNullServiceCharges();
-        serviceChargeRepository.deleteAllById(noReferenceChargeIds);
-        List<Long> noReferenceServiceIds = serviceRepository.findNullCategoryServices();
-        serviceRepository.deleteAllById(noReferenceServiceIds);
+        boolean isExist = categoryRepository.existsById(categoryId);
+        if(isExist){
+            category.setCategoryId(categoryId);
+            categoryRepository.save(category);
+            List<Long> noReferenceChargeIds = serviceChargeRepository.findNullServiceCharges();
+            serviceChargeRepository.deleteAllById(noReferenceChargeIds);
+            List<Long> noReferenceServiceIds = serviceRepository.findNullCategoryServices();
+            serviceRepository.deleteAllById(noReferenceServiceIds);
+            return true;
+        }else {
+            return false;
+        }
+
     }
 
     @Override
@@ -63,5 +72,18 @@ public class CategoryServiceImpl implements CategoryService{
     @Override
     public void updateCategoryIconById(Long categoryId, String categoryIconPath) {
         categoryRepository.updateCategoryIconByCategoryId(categoryIconPath, categoryId);
+    }
+    @Override
+    public ManageCategoryDTO getManageCategoryDetails(int pageNo, int noOfElements) {
+        Pageable pageableCriteria = PageRequest.of(pageNo, noOfElements, Sort.by("categoryName"));
+        Page<Category> categoryPage = categoryRepository.findAll(pageableCriteria);
+
+        List<ManageCategoryDetails> manageCategoryDetailsList = new ArrayList<>();
+        categoryPage.forEach(category -> manageCategoryDetailsList.add(new ManageCategoryDetails(category.getCategoryId(),category.getCategoryName(),category.getServiceList().size())));
+        ManageCategoryDTO manageCategoryDTO = new ManageCategoryDTO();
+        manageCategoryDTO.setPage(pageNo);
+        manageCategoryDTO.setNoOfElements(noOfElements);
+        manageCategoryDTO.setManageCategoryDetailsList(manageCategoryDetailsList);
+        return manageCategoryDTO;
     }
 }
